+++ frontend/src/hooks/useWorksheetData.js
@@ -1,5 +1,5 @@
-import { useState } from 'react'
-import { postWorkSheet, putWorkSheet } from '../api/WorkSheetAPI'
+import { useState, useEffect } from 'react'
+import { postWorkSheet, putWorkSheet, getWorkSheet } from '../api/WorkSheetAPI'
 import { useParams, useHistory } from 'react-router-dom'
 import { PATH_VARIABLES } from '../Const'
 
@@ -7,7 +7,9 @@ export default function useWorksheetData() {
   const [sent, setSent] = useState(false)
   const [sentSuccessfully, setSentSuccessfully] = useState(false)
   const [popUpMessage, setPopUpMessage] = useState('')
+  const [worksheetData, setWorksheetData] = useState()
   const { id } = useParams()
+  const [WorkSheetDataHandle, setWorkSheetDataHandle] = useState(true)
 
   const history = useHistory()
 
@@ -16,15 +18,30 @@ export default function useWorksheetData() {
     setSent(false)
   }
 
+  useEffect(() => {
+    if (id !== undefined) {
+      async function fetchData() {
+        try {
+          const response = await getWorkSheet(id)
+          setWorksheetData({ ...response.data, loaded: true })
+        } catch (error) {
+          setWorksheetData({ loaded: true })
+          setPopUpMessage('A módosítás sikertelen')
+          setSent(true)
+        }
+      }
+      fetchData()
+    }
+  }, [id])
+
   async function HandleData(values) {
-    const { WorkSheetDataHandle, setWorkSheetDataHandle } = useState()
-    if (id === null) {
-      setWorkSheetDataHandle(putWorkSheet(values))
-    } else {
-      setWorkSheetDataHandle(postWorkSheet(id, values))
+    if (id === undefined) {
+      setWorkSheetDataHandle(false)
     }
     try {
-      const response = await WorkSheetDataHandle
+      const response = WorkSheetDataHandle
+        ? await putWorkSheet(values)
+        : await postWorkSheet(id, values)
       if (response.status === 200) {
         setPopUpMessage('Munkalap sikeresen módosítva')
         setSentSuccessfully(true)
@@ -45,5 +62,6 @@ export default function useWorksheetData() {
     handleClick,
     popUpMessage,
     sent,
+    worksheetData,
   }
 }
